Move RankingsResponse type to shared module and simplify import paths
