+++ modules/uiTypes.ts
@@ -1,0 +1,1 @@
+export type RankingsResponse = any
